refactor: add routes auth and device for better structure
