+++ backend/src/types/types.ts
@@ -1,0 +1,4 @@
+export interface Env {
+  DATABASE_URL: string;
+  JWT_SECRET: string;
+}
